+++ proxy/post.js
@@ -1,4 +1,5 @@
 var Post = require('../models').Post;
+var Category = require('./category');
 
 module.exports.createNewPost = function (title, author, date, tags, post, category, callback) {
     var p = new Post();
@@ -8,6 +9,7 @@ module.exports.createNewPost = function (title, author, date, tags, post, catego
     p.tags = tags;
     p.post = post;
     p.category = category;
+    // Category.increaseCount(category);
     p.save(callback);
 };
 
@@ -20,7 +22,7 @@ module.exports.getOnePostById = function (id, callback) {
 };
 
 module.exports.getTenPosts = function (callback) {
-    Post.find({}, {limit: 10}, callback);
+    Post.find({}, null,{limit: 10}, callback);
 };
 
 
@@ -29,5 +31,21 @@ module.exports.getAllPosts = function (callback) {
 };
 
 module.exports.deletePost = function (id, callback) {
-    Post.findByIdAndRemove(id, callback);
+    Post.findOneById(id, function (err, p) {
+        if (err) {
+            callback(err);
+        }
+        if (p) {
+            // Category.decreaseCount(p.category);
+            Post.findByIdAndRemove(id, callback);
+        }
+    });
+};
+
+module.exports.getUserPost = function (user, callback) {
+    Post.find({author: user}, callback);
+};
+
+module.exports.getUserTenPosts = function (user, callback) {
+    Post.find({author: user}, null, {limit: 10}, callback);
 };
